Code refactoring to avoid duplication
